requirements.txt
Flask==1.1.1
mysql-connector==2.2.9
boto3==1.9.199
